+++ src/app/warning/directives.js
@@ -10,6 +10,22 @@ function warningPanelDirective() {
     };
 }
 
+function fileReadDirective() {
+    return {
+        scope: {
+            fileread: "="
+        },
+        link: function (scope, element, attributes) {
+            element.bind("change", function (changeEvent) {
+                scope.$apply(function () {
+                    scope.fileread = changeEvent.target.files[0];
+                });
+            });
+        }
+    }
+}
+
 module.exports = {
-    warningPanelDirective: warningPanelDirective
+    warningPanelDirective: warningPanelDirective,
+    fileReadDirective: fileReadDirective
 };