+++ CHANGES.md
@@ -1,10 +1,16 @@
 2.24.1dev0
-===================
+==========
 
 **Enhancements**
 
+* Add support for Suricate options in warning form
+* Add activity and magnitude problem fields in warning form
+* Get options list (magnitude problem, activities, categories) from endpoint
+**Warning:** these features are only available for Geotrek-admin >= 2.32.12
+
 **Bug fixes**
 
+* 
 
 2.24.0 / 2020-04-28
 ===================
@@ -17,6 +23,12 @@
     - add a text about personal data
 * Add translation file for portugese
 * Fix some translation error
+
+* Add support for Suricate options in warning form
+    - add activity and magnitude problem fields
+    - get options list from endpoint
+    - display activity and magnitude only if options list is available from API
+    **Warning:** this feature is only available for Geotrek-admin >= 2.32.12
 
 **Bug fixes**